+++ src/index.jsx
@@ -1,13 +1,16 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
+// import React from 'react'
+// import ReactDOM from 'react-dom'
 import App from './app/App'
-import { StoreContextProvider } from './store/StoreProvider'
+// import { StoreContextProvider } from './store/StoreProvider'
 
-ReactDOM.render(
-  <React.StrictMode>
-    <StoreContextProvider>
-      <App />
-    </StoreContextProvider>
-  </React.StrictMode>,
-  document.getElementById('root')
-)
+// ReactDOM.render(
+//   <React.StrictMode>
+//     <StoreContextProvider>
+//       <App />
+//     </StoreContextProvider>
+//   </React.StrictMode>,
+//   document.getElementById('root')
+// )
+
+console.log("INDEXJSX")
+console.log(App);
